\name{as.data.frame.correlation}
\alias{as.data.frame.correlation}
\title{as...}
\usage{as.data.frame.correlation(x, ...)}
\description{as.data.frame for correlation object.}
\details{as.data.frame for correlation object (internal).}
\author{David Hajage}
\keyword{internal}
\arguments{\item{x}{a correlation object}
\item{...}{not used}}
