\name{as.data.frame.freq}
\alias{as.data.frame.freq}
\title{as.data.frame for freq object.}
\usage{
  as.data.frame.freq(x, ...)
}
\arguments{
  \item{x}{a freq object}

  \item{...}{not used}
}
\description{
  as.data.frame for freq object (internal).
}
\author{
  David Hajage
}
\keyword{internal}

