\name{as.data.frame.summarize.by}
\alias{as.data.frame.summarize.by}
\title{as...}
\usage{as.data.frame.summarize.by(x, ...)}
\description{as.data.frame for summarize.by object.}
\details{as.data.frame for summarize.by object (internal).}
\author{David Hajage}
\keyword{internal}
\arguments{\item{x}{a summarize.by object}
\item{...}{not used}}
