\name{as.data.frame.summarize}
\alias{as.data.frame.summarize}
\title{as...}
\usage{as.data.frame.summarize(x, ...)}
\description{as.data.frame for summarize object.}
\details{as.data.frame for summarize object (internal).}
\author{David Hajage}
\keyword{internal}
\arguments{\item{x}{a summarize object}
\item{...}{not used}}
