\name{as.data.frame.survival}
\alias{as.data.frame.survival}
\title{as.data.frame for survival object.}
\usage{
  as.data.frame.survival(x, ...)
}
\arguments{
  \item{x}{a summarize object}

  \item{...}{not used}
}
\description{
  as.data.frame for survival object (internal).
}
\author{
  David Hajage
}
\keyword{internal}

