\name{as.data.frame.tabular}
\alias{as.data.frame.tabular}
\title{as...}
\usage{as.data.frame.tabular(x, ...)}
\description{as.data.frame for tabular object.}
\details{as.data.frame for tabular object (internal).}
\author{David Hajage}
\keyword{internal}
\arguments{\item{x}{a tabular object}
\item{...}{not used}}
