\name{ascii.correlation}
\alias{ascii.correlation}
\title{Ascii for correlation object.}
\usage{ascii.correlation(x, format="nice", digits=5, include.rownames=TRUE,
    include.colnames=TRUE, header=TRUE, ...)}
\description{Ascii for correlation object.}
\details{Ascii method for correlation object (internal).}
\author{David Hajage}
\keyword{internal}
\arguments{\item{x}{a correlation object}
\item{format}{see \code{?ascii} in \code{ascii} package}
\item{digits}{see \code{?ascii} in \code{ascii} package}
\item{include.rownames}{see \code{?ascii} in \code{ascii} package}
\item{include.colnames}{see \code{?ascii} in \code{ascii} package}
\item{header}{see \code{?ascii} in \code{ascii} package}
\item{...}{other arguments passed to \code{ascii}}}
