\name{ascii.remix}
\alias{ascii.remix}
\title{Ascii for remix object.}
\usage{
  \method{ascii}{remix} (x, caption.level = c("s", "e",
  "m"), format = "nice", digits = 2, ...)
}
\arguments{
  \item{x}{a remix object}

  \item{caption.level}{see \code{?ascii} in \code{ascii}
  package}

  \item{format}{see \code{?ascii} in \code{ascii} package}

  \item{digits}{see \code{?ascii} in \code{ascii} package}

  \item{...}{other arguments passed to \code{ascii} (all
  except \code{caption} which has no effect)}
}
\description{
  Ascii method for remix object.
}
\author{
  David Hajage
}
\keyword{univar}

