\name{ascii.summarize.by}
\alias{ascii.summarize.by}
\title{Ascii for summarize...}
\usage{ascii.summarize.by(x, format="nice", digits=5, include.rownames=FALSE,
    include.colnames=TRUE, header=TRUE, lgroup=attr(x, "lgroup"),
    n.lgroup=attr(x, "n.lgroup"), ...)}
\description{Ascii for summarize.by object.}
\details{Ascii method for summarize.by object (internal).}
\author{David Hajage}
\keyword{univar}
\arguments{\item{x}{a summarize.by object}
\item{format}{see \code{?ascii} in \code{ascii} package}
\item{digits}{see \code{?ascii} in \code{ascii} package}
\item{include.rownames}{see \code{?ascii} in \code{ascii} package}
\item{include.colnames}{see \code{?ascii} in \code{ascii} package}
\item{header}{see \code{?ascii} in \code{ascii} package}
\item{lgroup}{see \code{?ascii} in \code{ascii} package}
\item{n.lgroup}{see \code{?ascii} in \code{ascii} package}
\item{...}{other arguments passed to \code{ascii}}}
