\name{ascii.tabular}
\alias{ascii.tabular}
\title{Ascii for tabular object.}
\usage{ascii.tabular(x, format="nice", digits=5, include.rownames=FALSE,
    include.colnames=TRUE, header=TRUE, lgroup=attr(x, "lgroup"),
    n.lgroup=attr(x, "n.lgroup"), tgroup=attr(x, "tgroup"),
    n.tgroup=attr(x, "n.tgroup"), ...)}
\description{Ascii for tabular object.}
\details{Ascii method for tabular object (internal).}
\author{David Hajage}
\keyword{univar}
\arguments{\item{x}{a tabular object}
\item{format}{see \code{?ascii} in \code{ascii} package}
\item{digits}{see \code{?ascii} in \code{ascii} package}
\item{include.rownames}{see \code{?ascii} in \code{ascii} package}
\item{include.colnames}{see \code{?ascii} in \code{ascii} package}
\item{header}{see \code{?ascii} in \code{ascii} package}
\item{lgroup}{see \code{?ascii} in \code{ascii} package}
\item{n.lgroup}{see \code{?ascii} in \code{ascii} package}
\item{tgroup}{see \code{?ascii} in \code{ascii} package}
\item{n.tgroup}{see \code{?ascii} in \code{ascii} package}
\item{...}{other arguments passed to \code{ascii}}}
