\name{check_formula}
\alias{check_formula}
\title{Check if a variable is repeated several times in a formula...}
\usage{check_formula(formula)}
\description{Check if a variable is repeated several times in a formula}
\author{David Hajage}
\keyword{internal}
\arguments{\item{formula}{formula (character)}}
