\name{correlation.data.frame}
\alias{correlation.data.frame}
\title{Compute correlation (data...}
\usage{correlation.data.frame(dfx, dfy, method=c("pearson", "kendall",
    "spearman"))}
\description{Compute correlation (data.frame input)}
\author{David Hajage}
\keyword{internal}
\arguments{\item{dfx}{data.frame}
\item{dfy}{data.frame}
\item{method}{method}}
