\name{demix}
\alias{demix}
\title{Demix...}
\usage{demix(x)}
\description{Demix}
\details{Transfrom a remix object into a (list of) data.frame(s).}
\value{A list of data.frame.}
\author{David Hajage}
\seealso{\code{remix}}
\arguments{\item{x}{a remix object}}
\examples{x <- remix(... ~ ., esoph, cum = TRUE)
demix(x)}
