\name{freq.data.frame}
\alias{freq.data.frame}
\title{Compute frequencies (data...}
\usage{freq.data.frame(df, useNA=c("no", "ifany", "always"), cum=FALSE)}
\description{Compute frequencies (data.frame input)}
\author{David Hajage}
\keyword{internal}
\arguments{\item{df}{data.frame}
\item{useNA}{useNA}
\item{cum}{logical}}
