\name{freq}
\alias{freq}
\title{Compute frequencies}
\usage{
  freq(x, useNA = c("no", "ifany", "always"), propNA =
  TRUE, cum = FALSE, addmargins = FALSE)
}
\arguments{
  \item{x}{factor}

  \item{useNA}{useNA}

  \item{propNA}{propNA}

  \item{cum}{logical}

  \item{addmargins}{addmargins}
}
\description{
  Compute frequencies
}
\author{
  David Hajage
}
\keyword{internal}

