\name{n.table}
\alias{n.table}
\title{table with addmargins...}
\usage{n.table(x, y, useNA=c("no", "ifany", "always"), margin=0:2,
    addmargins=FALSE)}
\description{\code{table} with addmargins}
\author{David Hajage}
\keyword{internal}
\arguments{\item{...}{...}
\item{useNA}{useNA}
\item{margin}{margin}
\item{addmargins}{addmargins}}
