\name{p.table}
\alias{p.table}
\title{prop...}
\usage{p.table(x, y, useNA=c("no", "ifany", "always"), margin=0:2,
    addmargins=FALSE, propNA=TRUE)}
\description{\code{prop.table} with cell proportion, addmargins and propNA}
\author{David Hajage}
\keyword{internal}
\arguments{\item{...}{...}
\item{useNA}{useNA}
\item{margin}{margin}
\item{addmargins}{addmargins}
\item{propNA}{propNA}}
