\name{print.correlation}
\alias{print.correlation}
\title{Print correlation object.}
\usage{
  \method{print}{correlation} (x, type = "rest", ...)
}
\arguments{
  \item{x}{a correlation object}

  \item{type}{type of output (see \code{?ascii} in
  \code{ascii} package)}

  \item{...}{other arguments passed to \code{ascii}}
}
\description{
  Print correlation object (internal).
}
\author{
  David Hajage
}
\keyword{internal}

