\name{print.freq}
\alias{print.freq}
\title{Print freq object.}
\usage{
  \method{print}{freq} (x, type = "rest", lstyle = "", ...)
}
\arguments{
  \item{x}{a freq object}

  \item{type}{type of output (see \code{?ascii} in
  \code{ascii} package)}

  \item{lstyle}{see \code{?ascii} in \code{ascii} package}

  \item{...}{other arguments passed to \code{ascii}}
}
\description{
  Print freq object (internal).
}
\author{
  David Hajage
}
\keyword{univar}

