\name{print.remix}
\alias{print.remix}
\title{
  Print remix object.
}
\description{
  Print remix object.
}
\usage{
  print.remix(x, type = "rest", caption.level = 1:3, ...)
}
\arguments{
  \item{x}{
    A remix object.
  }
  \item{type}{
    type of output (see \code{?ascii} in \code{ascii} package)
  }
  \item{caption.level}{
    captions levels (see \code{?ascii} in \code{ascii} package)
  }
  \item{\dots}{
    Not used
  }
}

\author{
David Hajage
}

\seealso{
  \code{\link{remix}}
}

\keyword{univar}

