\name{print.resume}
\alias{print.resume}
\alias{print.resume.by}
\title{
  Print resume object.
}
\description{
  Print resume object.
}
\usage{
print.resume(x, ...)
print.resume.by(x, ...)
}
\arguments{
  \item{x}{
    A resume object.
  }
  \item{\dots}{
    Not used
  }
}

\author{
David Hajage
}

\seealso{
  \code{\link{resume}}
}

\keyword{univar}

