\name{print.summarize.by}
\alias{print.summarize.by}
\title{Print summarize...}
\usage{print.summarize.by(x, type="rest", lstyle="", ...)}
\description{Print summarize.by object.}
\details{Print summarize.by object (internal).}
\author{David Hajage}
\keyword{internal}
\arguments{\item{x}{a summarize.by object}
\item{type}{type of output (see \code{?ascii} in \code{ascii} package)}
\item{...}{other arguments passed to \code{ascii}}}
