\name{print.summarize}
\alias{print.summarize}
\title{Print summarize object.}
\usage{print.summarize(x, type="rest", lstyle="", ...)}
\description{Print summarize object.}
\details{Print summarize object (internal).}
\author{David Hajage}
\keyword{internal}
\arguments{\item{x}{a summarize object}
\item{type}{type of output (see \code{?ascii} in \code{ascii}
package)}
\item{...}{other arguments passed to \code{ascii}}}
