\name{print.survival}
\alias{print.survival}
\title{Print survival object.}
\usage{
  \method{print}{survival} (x, type = "rest", lstyle = "",
  ...)
}
\arguments{
  \item{x}{a summarize object}

  \item{type}{type of output (see \code{?ascii} in
  \code{ascii} package)}

  \item{lstyle}{see \code{?ascii} in \code{ascii} package}

  \item{...}{other arguments passed to \code{ascii}}
}
\description{
  Print summarize object (internal).
}
\author{
  David Hajage
}
\keyword{internal}

