\name{print.tabular}
\alias{print.tabular}
\title{
  Print tabular object.
}
\description{
  Print tabular object.
}
\usage{
print.tabular(x, ...)
}
\arguments{
  \item{x}{
    A tabular object.
  }
  \item{\dots}{
    Not used
  }
}

\author{
David Hajage
}

\seealso{
  \code{\link{tabular}}
}

\keyword{univar}

