\name{prop.table2}
\alias{prop.table2}
\title{prop...}
\usage{prop.table2(..., useNA=c("no", "ifany", "always"), margin=0,
    propNA=TRUE)}
\description{\code{prop.table} with cell and propNA}
\author{David Hajage}
\keyword{internal}
\arguments{\item{...}{...}
\item{useNA}{useNA}
\item{margin}{margin}
\item{propNA}{propNA}}
