\name{regroup}
\alias{regroup}
\title{Regroup factors with factors, and numerical variables with numerical variables...}
\usage{regroup(vars, numdata)}
\description{Regroup factors with factors, and numerical variables with numerical variables}
\author{David Hajage}
\keyword{internal}
\arguments{\item{vars}{vars}
\item{numdata}{numdata}}
