\name{summarize.data.frame}
\alias{summarize.data.frame}
\title{Compute summary statistics (data.frame input)}
\usage{
  summarize.data.frame(df, funs = c(mean, sd, quantile, n,
  na), label = FALSE, ...)
}
\arguments{
  \item{df}{a data.frame}

  \item{funs}{functions}

  \item{...}{passed to funs}
}
\description{
  Compute summary statistics (data.frame input)
}
\author{
  David Hajage
}
\keyword{internal}

