\name{summarize}
\alias{summarize}
\title{Compute summary statistics}
\usage{
  summarize(x, funs = c(mean, sd, quantile, n, na), ...)
}
\arguments{
  \item{x}{numeric object}

  \item{funs}{functions}

  \item{...}{passed to funs}
}
\description{
  Compute summary statistics
}
\author{
  David Hajage
}
\keyword{internal}

