\name{survival.data.frame}
\alias{survival.data.frame}
\title{Compute survival according to a factor}
\usage{
  survival.data.frame(df, times = NULL, label = FALSE)
}
\arguments{
  \item{df}{df}

  \item{times}{times}

  \item{label}{label}
}
\description{
  Compute survival according to a factor
}
\author{
  David Hajage
}

