\name{survival}
\alias{survival}
\title{Compute survival}
\usage{
  survival(surv, by = NULL, times = NULL, followup = FALSE,
  test = FALSE, test.survival = test.survival.logrank,
  show.test = display.test, plim = 4, show.method = TRUE)
}
\arguments{
  \item{surv}{a Surv object}

  \item{by}{by}

  \item{times}{times}

  \item{followup}{followup}

  \item{test}{test}

  \item{test.survival}{test.survival}

  \item{show.test}{show.test}

  \item{plim}{plim}

  \item{show.method}{show.method}
}
\description{
  Compute survival
}
\author{
  David Hajage
}
\keyword{internal}

