\name{tabular.data.frame}
\alias{tabular.data.frame}
\title{Compute a contingency table (data...}
\usage{tabular.data.frame(dfx, dfy, margin=0:2, useNA=c("no", "ifany",
    "always"))}
\description{Compute a contingency table (data.frame input)}
\author{David Hajage}
\keyword{internal}
\arguments{\item{dfx}{data.frame}
\item{dfy}{data.frame}
\item{margin}{margin}
\item{useNA}{useNA}}
