\name{tabular}
\alias{tabular}
\title{Compute a contingency table...}
\usage{tabular(x, y, margin=0:2, useNA=c("no", "ifany", "always"))}
\description{Compute a contingency table}
\author{David Hajage}
\keyword{internal}
\arguments{\item{x}{factor}
\item{y}{factor}
\item{margin}{margin}
\item{useNA}{useNA}}
