\name{test.summarize.kruskal}
\alias{test.summarize.kruskal}
\title{test for mean comparison}
\usage{
  test.summarize.kruskal(x, g)
}
\arguments{
  \item{x}{vector}

  \item{g}{another vector}
}
\value{
  a list with two componments: p and method
}
\description{
  Compute a kruskal.test.
}
\author{
  David Hajage
}

