\name{test.tabular.fisher}
\alias{test.tabular.fisher}
\title{test for contingency table}
\usage{
  test.tabular.fisher(x, y, na = FALSE)
}
\arguments{
  \item{x}{vector}

  \item{y}{another vector}

  \item{na}{use NA ?}
}
\value{
  a list with two componments: p and method
}
\description{
  Compute a fisher test
}
\author{
  David Hajage
}

