\name{as.data.frame.correlation}
\alias{as.data.frame.correlation}
\title{as.data.frame for correlation object.}
\usage{
  as.data.frame.correlation(x, ...)
}
\arguments{
  \item{x}{a correlation object}

  \item{...}{not used}
}
\description{
  as.data.frame for correlation object (internal).
}
\author{
  David Hajage
}
\keyword{internal}

