\name{as.data.frame.summarize.by}
\alias{as.data.frame.summarize.by}
\title{as.data.frame for summarize.by object.}
\usage{
  as.data.frame.summarize.by(x, ...)
}
\arguments{
  \item{x}{a summarize.by object}

  \item{...}{not used}
}
\description{
  as.data.frame for summarize.by object (internal).
}
\author{
  David Hajage
}
\keyword{internal}

