\name{as.data.frame.summarize}
\alias{as.data.frame.summarize}
\title{as.data.frame for summarize object.}
\usage{
  as.data.frame.summarize(x, ...)
}
\arguments{
  \item{x}{a summarize object}

  \item{...}{not used}
}
\description{
  as.data.frame for summarize object (internal).
}
\author{
  David Hajage
}
\keyword{internal}

