\name{as.data.frame.tabular}
\alias{as.data.frame.tabular}
\title{as.data.frame for tabular object.}
\usage{
  as.data.frame.tabular(x, ...)
}
\arguments{
  \item{x}{a tabular object}

  \item{...}{not used}
}
\description{
  as.data.frame for tabular object (internal).
}
\author{
  David Hajage
}
\keyword{internal}

