\name{ascii.freq}
\alias{ascii.freq}
\title{Ascii for freq object.}
\usage{
  \method{ascii}{freq} (x, format = "nice", digits = 3,
  include.rownames = FALSE, include.colnames = TRUE, header
  = TRUE, lgroup = attr(x, "lgroup"), n.lgroup = attr(x,
  "n.lgroup"), ...)
}
\arguments{
  \item{x}{a freq object}

  \item{format}{see \code{?ascii} in \code{ascii} package}

  \item{digits}{see \code{?ascii} in \code{ascii} package}

  \item{include.rownames}{see \code{?ascii} in \code{ascii}
  package}

  \item{rownames}{see \code{?ascii} in \code{ascii}
  package}

  \item{include.colnames}{see \code{?ascii} in \code{ascii}
  package}

  \item{header}{see \code{?ascii} in \code{ascii} package}

  \item{lgroup}{see \code{?ascii} in \code{ascii} package}

  \item{n.lgroup}{see \code{?ascii} in \code{ascii}
  package}

  \item{...}{other arguments passed to \code{ascii}}
}
\description{
  Ascii method for freq object (internal).
}
\author{
  David Hajage
}
\keyword{univar}

