\name{ascii.summarize.by}
\alias{ascii.summarize.by}
\title{Ascii for summarize.by object.}
\usage{
  \method{ascii}{summarize.by} (x, format = "nice", digits
  = 5, include.rownames = FALSE, include.colnames = TRUE,
  header = TRUE, lgroup = attr(x, "lgroup"), n.lgroup =
  attr(x, "n.lgroup"), rgroup = attr(x, "rgroup"), n.rgroup
  = attr(x, "n.rgroup"), rstyle = "d", ...)
}
\arguments{
  \item{x}{a summarize.by object}

  \item{format}{see \code{?ascii} in \code{ascii} package}

  \item{digits}{see \code{?ascii} in \code{ascii} package}

  \item{include.rownames}{see \code{?ascii} in \code{ascii}
  package}

  \item{include.colnames}{see \code{?ascii} in \code{ascii}
  package}

  \item{header}{see \code{?ascii} in \code{ascii} package}

  \item{lgroup}{see \code{?ascii} in \code{ascii} package}

  \item{n.lgroup}{see \code{?ascii} in \code{ascii}
  package}

  \item{rgroup}{see \code{?ascii} in \code{ascii} package}

  \item{n.rgroup}{see \code{?ascii} in \code{ascii}
  package}

  \item{rstyle}{see \code{?ascii} in \code{ascii} package}

  \item{...}{other arguments passed to \code{ascii}}
}
\description{
  Ascii method for summarize.by object (internal).
}
\author{
  David Hajage
}
\keyword{univar}

