\name{ascii.tabular}
\alias{ascii.tabular}
\title{Ascii for tabular object.}
\usage{
  \method{ascii}{tabular} (x, format = "nice", digits = 5,
  include.rownames = FALSE, include.colnames = TRUE, header
  = TRUE, rstyle = "d", caption = NULL, caption.level =
  NULL, ...)
}
\arguments{
  \item{x}{a tabular object}

  \item{format}{see \code{?ascii} in \code{ascii} package}

  \item{digits}{see \code{?ascii} in \code{ascii} package}

  \item{include.rownames}{see \code{?ascii} in \code{ascii}
  package}

  \item{include.colnames}{see \code{?ascii} in \code{ascii}
  package}

  \item{header}{see \code{?ascii} in \code{ascii} package}

  \item{rstyle}{see \code{?ascii} in \code{ascii} package}

  \item{caption}{see \code{?ascii} in \code{ascii} package}

  \item{caption.level}{see \code{?ascii} in \code{ascii}
  package}

  \item{...}{other arguments passed to \code{ascii}}
}
\description{
  Ascii method for tabular object (internal).
}
\author{
  David Hajage
}
\keyword{univar}

