\name{correlation.data.frame}
\alias{correlation.data.frame}
\title{Compute correlation (data.frame input)}
\usage{
  correlation.data.frame(dfx, dfy, method = c("pearson",
  "kendall", "spearman"))
}
\arguments{
  \item{dfx}{data.frame}

  \item{dfy}{data.frame}

  \item{method}{method}
}
\description{
  Compute correlation (data.frame input)
}
\author{
  David Hajage
}
\keyword{internal}

