\name{correlation}
\alias{correlation}
\title{Compute correlation}
\usage{
  correlation(x, y, method = c("pearson", "kendall",
  "spearman"))
}
\arguments{
  \item{x}{numerical object}

  \item{y}{numerical object}

  \item{method}{method}
}
\description{
  Compute correlation
}
\author{
  David Hajage
}
\keyword{internal}

