\name{freq.data.frame}
\alias{freq.data.frame}
\title{Compute frequencies (data.frame input)}
\usage{
  freq.data.frame(df, useNA = c("no", "ifany", "always"),
  propNA = TRUE, cum = FALSE, addmargins = FALSE, label =
  FALSE)
}
\arguments{
  \item{df}{data.frame}

  \item{useNA}{useNA}

  \item{propNA}{propNA}

  \item{cum}{logical}
}
\description{
  Compute frequencies (data.frame input)
}
\author{
  David Hajage
}
\keyword{internal}

