\name{n.table}
\alias{n.table}
\title{\code{table} with addmargins and test}
\usage{
  n.table(x, y = NULL, useNA = c("no", "ifany", "always"),
  margin = 0:2, addmargins = FALSE, test = FALSE,
  test.tabular = test.tabular.auto, show.test =
  display.test, plim = 4, show.method = TRUE, na = FALSE)
}
\arguments{
  \item{x}{x}

  \item{y}{y}

  \item{useNA}{useNA}

  \item{margin}{margin}

  \item{addmargins}{addmargins}
}
\description{
  \code{table} with addmargins and test
}
\author{
  David Hajage
}
\keyword{internal}

