\name{p.table}
\alias{p.table}
\title{\code{prop.table} with cell proportion, addmargins and propNA}
\usage{
  p.table(x, y = NULL, useNA = c("no", "ifany", "always"),
  margin = 0:2, addmargins = FALSE, propNA = TRUE)
}
\arguments{
  \item{...}{...}

  \item{useNA}{useNA}

  \item{margin}{margin}

  \item{addmargins}{addmargins}

  \item{propNA}{propNA}
}
\description{
  \code{prop.table} with cell proportion, addmargins and
  propNA
}
\author{
  David Hajage
}
\keyword{internal}

