\name{print.correlation}
\alias{print.correlation}
\title{Print correlation object.}
\usage{print.correlation(x, type="rest", ...)}
\description{Print correlation object.}
\details{Print correlation object (internal).}
\author{David Hajage}
\keyword{internal}
\arguments{\item{x}{a correlation object}
\item{type}{type of output (see \code{?ascii} in \code{ascii}
package)}
\item{...}{other arguments passed to \code{ascii}}}
