\name{print.correlation}
\alias{print.correlation}
\title{
  Print correlation object.
}
\description{
  Print correlation object (internal).
}
\usage{
  print.correlation(x, type = "rest", ...)
}
\arguments{
  \item{x}{
    A correlation object.
  }
  \item{type}{
    type of output (see \code{?ascii} in \code{ascii} package)
  }
  \item{\dots}{
    Not used
  }
}

\author{
David Hajage
}

\keyword{univar}

