\name{print.correlation}
\alias{print.correlation}
\title{
  Print correlation object.
}
\description{
  Print correlation object.
}
\usage{
print.correlation(x, ...)
}
\arguments{
  \item{x}{
    A correlation object.
  }
  \item{\dots}{
    Not used
  }
}

\author{
David Hajage
}

\seealso{
  \code{\link{correlation}}
}

\keyword{univar}

