\name{print.freq}
\alias{print.freq}
\title{
  Print freq object.
}
\description{
  Print freq object (internal).
}
\usage{
print.freq(x, type = "rest", ...)
}
\arguments{
  \item{x}{
    A freq object.
  }
  \item{type}{
    type of output (see \code{?ascii} in \code{ascii} package)
  }
  \item{\dots}{
    Not used
  }
}

\author{
David Hajage
}

\keyword{univar}

