\name{print.freq}
\alias{print.freq}
\title{
  Print freq object.
}
\description{
  Print freq object.
}
\usage{
print.freq(x, ...)
}
\arguments{
  \item{x}{
    A freq object.
  }
  \item{\dots}{
    Not used
  }
}

\author{
David Hajage
}

\seealso{
  \code{\link{freq}}
}

\keyword{univar}

