\name{print.freq}
\alias{print.freq}
\title{Print freq object.}
\usage{print.freq(x, type="rest", ...)}
\description{Print freq object.}
\details{Print freq object (internal).}
\author{David Hajage}
\arguments{\item{x}{a freq object}
\item{type}{type of output (see \code{?ascii} in \code{ascii} package)}
\item{\dots}{other arguments passed to \code{ascii}}}
