\name{print.remix}
\alias{print.remix}
\title{Print a remix object...}
\usage{print.remix(x, type="rest", caption.level=1:3, lstyle="", tstyle="",
    ...)}
\description{Print a remix object}
\details{Print remix object using ascii package}
\author{David Hajage}
\keyword{univar}
\arguments{\item{x}{a remix object}
\item{type}{type of output. See \code{?ascii} in \code{ascii} package}
\item{caption.level}{see \code{?ascii} in \code{ascii} package}
\item{lstyle}{see \code{?ascii} in \code{ascii} package}
\item{tstyle}{see \code{?ascii} in \code{ascii} package}
\item{...}{other arguments passed to \code{ascii} (all except \code{caption}
which has no effect)}}
