\name{print.remix}
\alias{print.remix}
\title{
  Print remix object.
}
\description{
  Print remix object.
}
\usage{
print.remix(x, ...)
}
\arguments{
  \item{x}{
    A remix object.
  }
  \item{\dots}{
    Not used
  }
}

\author{
David Hajage
}

\seealso{
  \code{\link{remix}}
}

\keyword{univar}

