\name{print.resume}
\alias{print.resume}
\alias{print.resume.by}
\title{
  Print resume object.
}
\description{
  Print resume object (internal).
}
\usage{
print.resume(x, type = "rest", ...)
print.resume.by(x, type = "rest", ...)
}
\arguments{
  \item{x}{
    A resume object.
  }
  \item{type}{
    type of output (see \code{?ascii} in \code{ascii} package)
  }
  \item{\dots}{
    Not used
  }
}

\author{
David Hajage
}

\keyword{univar}

