\name{print.summarize.by}
\alias{print.summarize.by}
\title{Print summarize.by object.}
\usage{
  \method{print}{summarize.by} (x, type = "rest", lstyle =
  "", ...)
}
\arguments{
  \item{x}{a summarize.by object}

  \item{type}{type of output (see \code{?ascii} in
  \code{ascii} package)}

  \item{...}{other arguments passed to \code{ascii}}
}
\description{
  Print summarize.by object (internal).
}
\author{
  David Hajage
}
\keyword{internal}

