\name{print.tabular}
\alias{print.tabular}
\title{Print tabular object.}
\usage{print.tabular(x, type="rest", ...)}
\description{Print tabular object.}
\details{Print tabular object (internal).}
\author{David Hajage}
\arguments{\item{x}{a tabular object}
\item{type}{type of output (see \code{?ascii} in \code{ascii} package)}
\item{\dots}{other arguments passed to \code{ascii}}}
