\name{print.tabular}
\alias{print.tabular}
\title{Print tabular object.}
\usage{
  \method{print}{tabular} (x, type = "rest", lstyle = "",
  tstyle = "", ...)
}
\arguments{
  \item{x}{a tabular object}

  \item{type}{type of output (see \code{?ascii} in
  \code{ascii} package)}

  \item{lstyle}{see \code{?ascii} in \code{ascii} package}

  \item{tstyle}{see \code{?ascii} in \code{ascii} package}

  \item{...}{other arguments passed to \code{ascii}}
}
\description{
  Print tabular object (internal).
}
\author{
  David Hajage
}
\keyword{univar}

