\name{print.tabular}
\alias{print.tabular}
\title{
  Print tabular object.
}
\description{
  Print tabular object (internal).
}
\usage{
print.tabular(x, type = "rest", ...)
}
\arguments{
  \item{x}{
    A tabular object.
  }
  \item{type}{
    type of output (see \code{?ascii} in \code{ascii} package)
  }
  \item{\dots}{
    Not used
  }
}

\author{
David Hajage
}

\keyword{univar}

