\name{survival.data.frame}
\alias{survival.data.frame}
\title{Compute survival (data.frame input)}
\usage{
  survival.data.frame(df, times = NULL, followup = FALSE,
  label = FALSE)
}
\arguments{
  \item{df}{df}

  \item{times}{times}

  \item{followup}{followup}

  \item{label}{label}
}
\description{
  Compute survival (data.frame input)
}
\author{
  David Hajage
}

