\name{tabular.data.frame}
\alias{tabular.data.frame}
\title{Compute a contingency table (data.frame input)}
\usage{
  tabular.data.frame(dfx, dfy, margin = 0:2, useNA =
  c("no", "ifany", "always"), propNA = TRUE, addmargins =
  FALSE, test = FALSE, test.tabular = test.tabular.auto,
  show.test = display.test, plim = 4, show.method = TRUE,
  label = FALSE)
}
\arguments{
  \item{dfx}{data.frame}

  \item{dfy}{data.frame}

  \item{margin}{margin}

  \item{useNA}{useNA}

  \item{propNA}{propNA}
}
\description{
  Compute a contingency table (data.frame input)
}
\author{
  David Hajage
}
\keyword{internal}

