\name{tabular}
\alias{tabular}
\title{##' Compute a contingency table...}
\usage{tabular(x, y, margin=0:2, useNA=c("no", "ifany", "always"),
    propNA=TRUE, addmargins=FALSE)}
\description{##' Compute a contingency table}
\author{David Hajage}
\keyword{internal}
\arguments{\item{x}{factor}
\item{y}{factor}
\item{margin}{margin}
\item{useNA}{useNA}
\item{propNA}{propNA}
\item{addmargins}{addmargins}}
