\name{tabular}
\alias{tabular}
\title{Compute a contingency table}
\usage{
  tabular(x, y, margin = 0:2, useNA = c("no", "ifany",
  "always"), propNA = TRUE, addmargins = FALSE, test =
  FALSE, test.tabular = test.tabular.auto, show.test =
  display.test, plim = 4, show.method = TRUE)
}
\arguments{
  \item{x}{factor}

  \item{y}{factor}

  \item{margin}{margin}

  \item{useNA}{useNA}

  \item{propNA}{propNA}

  \item{addmargins}{addmargins}
}
\description{
  Compute a contingency table
}
\author{
  David Hajage
}
\keyword{internal}

